[arch] use u64 as GuestAddress inner type

+++ arch/src/aarch64/layout.rs
@@ -50,11 +50,11 @@
 // Taken from (http://infocenter.arm.com/help/topic/com.arm.doc.den0001c/DEN0001C_principles_of_arm_memory_maps.pdf).
 
 /// Start of RAM on 64 bit ARM.
-pub const DRAM_MEM_START: usize = 0x8000_0000; // 2 GB.
+pub const DRAM_MEM_START: u64 = 0x8000_0000; // 2 GB.
 /// The maximum addressable RAM address.
-pub const DRAM_MEM_END: usize = 0x00FF_8000_0000; // 1024 - 2 = 1022 GB.
+pub const DRAM_MEM_END: u64 = 0x00FF_8000_0000; // 1024 - 2 = 1022 GB.
 /// The maximum RAM size.
-pub const DRAM_MEM_MAX_SIZE: usize = DRAM_MEM_END - DRAM_MEM_START;
+pub const DRAM_MEM_MAX_SIZE: u64 = DRAM_MEM_END - DRAM_MEM_START;
 
 /// Kernel command line maximum size.
 /// As per `arch/arm64/include/uapi/asm/setup.h`.
